% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arith-timeDay.R
\docType{methods}
\name{Arith-timeDay}
\alias{Arith-timeDay}
\alias{Arith,timeDay,POSIXt-method}
\alias{Arith,POSIXt,timeDay-method}
\alias{Arith,timeDay,Date-method}
\alias{Arith,Date,timeDay-method}
\title{Arithmetic Operators for \code{timeDay} objects}
\usage{
\S4method{Arith}{timeDay,POSIXt}(e1, e2)

\S4method{Arith}{POSIXt,timeDay}(e1, e2)

\S4method{Arith}{timeDay,Date}(e1, e2)

\S4method{Arith}{Date,timeDay}(e1, e2)
}
\arguments{
\item{e1, e2}{timeDay and POSIXt or Date objects. Missing values are permitted in 
either argument and result in a missing value in the output.}
}
\value{
A vector of class "POSIXct."
}
\description{
Addition of time-of-day data to either "Date" or "POSIXt" classes. 
This is useful when dates and times are recorded in separate
columns in a dataset.
}
\examples{
as.Date("2001-03-04") + as.timeDay("10:00")
\dontrun{
library(smwrData)
data(QW05078470)
# Note that the result is reported in the local time zone!
QW05078470$DATES + as.timeDay(QW05078470$TIMES)
}
}
\keyword{manip}
\keyword{methods}
