% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{as.timeDay}
\alias{as.timeDay}
\alias{as.timeDay,character,character-method}
\alias{as.timeDay,character,missing-method}
\alias{as.timeDay,numeric,missing-method}
\alias{as.timeDay,timeDay,missing-method}
\title{Methods for Function \code{as.timeDay}}
\usage{
as.timeDay(time, format)

\S4method{as.timeDay}{timeDay,missing}(time, format)

\S4method{as.timeDay}{numeric,missing}(time, format)

\S4method{as.timeDay}{character,character}(time, format)

\S4method{as.timeDay}{character,missing}(time, format)
}
\arguments{
\item{time}{the time of day in character format.}

\item{format}{The format for converting the time of day. See \code{\link{strptime}}
for possible format specifiers.}
}
\description{
Valid conversions for function \code{as.timeDay}.
}
\details{
Inconsistent formats for \code{time} will result in an error. Missing values in
\code{time} will result in missing values in the output.
}
\examples{
as.timeDay("10:00")
as.timeDay("3 PM", format="\%I \%p")
}
\keyword{manip}
\keyword{methods}

