% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxCox.R
\name{boxCox}
\alias{IboxCox}
\alias{boxCox}
\title{Box-Cox Power Transform}
\usage{
boxCox(x, lambda = 1, GM, alpha = 0)

IboxCox(x, lambda = 1, GM, alpha = 0)
}
\arguments{
\item{x}{a numeric vector to be transformed by \code{boxCox} or
back-transformed by \code{IboxCox}. Must be strictly positive for the
forward trasnformation---the argument
\code{alpha} can be used to force positive values. Missing values are
allowed and result in corresponding missing values in the output. See \bold{Details}.}

\item{lambda}{the power term in the Box-Cox transformation. The value of 1 is
a linear transform, the value of 0 results in a natural log transform.}

\item{GM}{the value to use for the geometric mean of \code{x}. If not
supplied, then compute the geometric mean (\code{boxCox}) or extract from
the attributes of \code{x} (\code{IboxCox}).}

\item{alpha}{an offset value for \code{x}.}
}
\value{
A numeric vector of the transformed or back-transformed values in
\code{x} with an attribute "GM" of the geometric mean.
}
\description{
Functions for transforming and back-transforming data using the Box-Cox power
transform, with options to preserve the measurement units.
}
\details{
If \code{x} contains missing values, then \code{GM} is computed after
omitting the missing values and the output vector has a missing value
wherever \code{x} has a missing value.\cr

The function \code{boxCox} computes the forward transform and the function
\code{IboxCox} computes the inverse [boxCox] transform, or back-transform.
}
\note{
The original power transform described by Box and Cox (1964) is
adjusted by a power transform of the geometric mean to retain the correct
dimensional units of the original data as described in section 13.2 by 
Draper and Smith (1998).
}
\examples{
X.test <- c(1,4,9,16,25,36,49)
boxCox(X.test)
boxCox(X.test, lambda=0)
IboxCox(boxCox(1:3, lambda=0), lambda=0) # verify the back-transform
## Should return
# [1] 1 2 3
# attr(,"GM")
# [1] 1.817121
}
\references{
Box, G.E.P., and Cox, D.R., 1964, An analysis of transformations:
Journal of the Royal Statistical Society, v. 26, Series B, p. 211--243.\cr
Draper, N.R., and Smith, H., 1998, Applied regression analysis: New York,
John Wiley and Sons, 706 p.
}
\seealso{
\code{\link{hyperbolic}}
}
\keyword{manip}

