\name{conc.meq}
\alias{conc.meq}
\title{Special Data}
\usage{
  conc.meq()
}
\value{
  A list containing necessary information for the function
  \code{con2meq}.
}
\description{
  Function to supply data for functions within USGSwsBase
}
\note{
  The user may choose to make local copies of data with the
  same name to change or add to each list.
}
\seealso{
  \code{\link{conc2meq}}
}
\keyword{list}

