% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daysInMonth.R
\name{daysInMonth}
\alias{daysInMonth}
\title{Days in a Month}
\usage{
daysInMonth(month, year, cum = FALSE)
}
\arguments{
\item{month}{the month number, must range in value from 1 to 12. Missing
values are permitted.}

\item{year}{the calendar year, replicated in length to match \code{month}.
Missing values are permitted.}

\item{cum}{a logical value. If \code{TRUE}, then return the cumulative days during the calendar year
at the end of each month. If \code{FALSE}, then return the number of days in the month.}
}
\value{
A vector matching \code{month} of the requested number of days.
Missing values are returned wherever either \code{month} or \code{year} is
missing.
}
\description{
Computes the number of days in a month or the total number of days in the
year to the end of the month.
}
\examples{
# Check February on a leap year and regular year.
# Should return 29, 28
daysInMonth(c(2,2), c(2000, 2001))
}
\keyword{manip}

