% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportCSV.R
\name{exportCSV}
\alias{exportCSV}
\title{Export Data}
\usage{
exportCSV(x, file.name = "")
}
\arguments{
\item{x}{the data frame to be written.}

\item{file.name}{a character string naming the file for output.}
}
\value{
The file name is returned.
}
\description{
Exports a data frame to a text-based file.
}
\note{
The function \code{exportCSV} also writes a meta file that contains
information about column formatting.
}
\seealso{
\code{\link[utils]{write.table}}, \code{\link{importCSV}}
}
\keyword{IO}
\keyword{manip}

