% Generated by roxygen2 (4.0.2): do not edit by hand
\name{findSites}
\alias{findSites}
\title{Site Inventory}
\usage{
findSites(type = c("SW", "ST", "GW", "WE", "SP", "LK", "ES", "OC"),
  state = "", county = "", huc = "", box = list(), name = "",
  data.type = c("any", "iv", "dv", "gw", "qw"))
}
\arguments{
\item{type}{the type of site desired. Must by one of "SW" or "ST" for
stream, "GW" for groundwater well, "WE" for wetland,
"SP" for spring, "LK" for lake, "ES" for esturary, or "OC" for ocean.}

\item{state}{the 2-character postal abbreviation for a state}

\item{county}{a character vector of county 5-digit FIPS ids. No
more than 20 can be selected. See \bold{References} to find county codes.}

\item{huc}{either a single 2-, 4-, or 6-digit hydrologic region, subregion,
or accounting unit, or from 1 to a maximum of 10 8-digit hydrologic
cataloging units.}

\item{box}{a list containing the components north, south, east, and west
that indicate the latitude and longitude extremes of the bounding box
for the retrieval. The components may be abbreviated n, s, e, and w.
The east and west values are assumed to be in terms of western
longitude, so may be either positive or negative values.}

\item{name}{the first few words of the name of the site. Most often used to search for
streamgages along a specific river. See \bold{Examples}.}

\item{data.type}{the type of data for which to search. Must be one of
"any" for any type of data, "iv" for instantaneous values, "dv" for daily
values, "gw" for ground-water levels, or "qw" for water-quality data.}
}
\value{
A data frame containing the columns.
#'\tabular{ll}{ agency_cd \tab The agency code\cr
site_no \tab The site identifier\cr
station_nm \tab The site name\cr
site_tp_cd \tab The site type code\cr
dec_lat_va \tab The decimal latitute\cr
dec_long_va \tab The decimal longitude\cr
coord_acy_cd \tab Latitude-longitude accuracy\cr
dec_coord_datum_cd \tab Decimal Latitude-longitude datum\cr
alt_va \tab Altitude of Gage/land surface\cr
alt_acy_va \tab Altitude accuracy\cr
alt_datum_cd \tab Altitude datum\cr
huc_cd \tab Hydrologic unit code\cr }
}
\description{
Get USGS streamgages, observation wells or other data collection sites
given some search criteria.
}
\details{
There are several arguments to the \code{findSites} function. Only one of
\code{state}, \code{county}, \code{huc} or \code{box} can be used to
define the major filter.
In general, only a few need
to be used to limit the search to a reasonable group.
}
\examples{
\dontrun{
# Find Streamgages on the Red River of the North
findSites(huc="0902", name="Red River", data.type="dv")
# Find wells in Greeley county Nebraska that have water-quality samples
findSites(type="GW", county="31077", data.type="qw")
}
}
\references{
Information about current water conditions in the United States
and historical hydrologic data can be obtained from
\url{http://waterdata.usgs.gov}.\cr
County FIPS codes can be obtained from
\url{http://www.epa.gov/enviro/html/codes/state.html}.\cr
}
\seealso{
\code{\link{readNWIS}}
}
\keyword{DataIO}

