% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hyperbolic}
\alias{Ihyperbolic}
\alias{hyperbolic}
\title{Hyperbolic transform}
\usage{
hyperbolic(x, factor = 0, scale = mean(x, na.rm = TRUE))

Ihyperbolic(x, factor = 0, scale)
}
\arguments{
\item{x}{a numeric vector to be transformed by \code{hyperbolic} or
back-trasnformed by \code{Ihyperbolic}. Must be strictly positive. Missing
values are allowed. See \bold{Details}.}

\item{factor}{the hyperbolic adjustment term in the hyperbolic equation.}

\item{scale}{the scaling factor for the data.}
}
\value{
A numeric vector of the transformed or back-transformed values in
\code{x} with an attribute "scale" of the values used for \code{scale}.
}
\description{
Functions for transforming and back-transforming data using a hyperbolic
function.
}
\details{
If \code{x} contains missing values, then \code{scale} is computed after
omitting the missing values and the output vector has a missing value
wherever \code{x} has a missing value.\cr

The basic equation for the hyberbolic transform is 1/(1 + (10^factor * x)/
scale). It is adjusted to produce fairly consistent values for small changes
in \code{factor}.\cr

The function \code{hyperbolic} computes the forward transform and the
function \code{Ihyperbolic} computes the inverse [hyperbolic] transform, or back-transform.
}
\note{
The original hyperbolic transform used a linear factor. The version in
these functions uses the common log of the factor to make the factors easier
to use.\cr

When used with the default value for \code{scale}, \code{factor} values
outside the range of +/- 3 have very little effect on the transform.
}
\examples{
hyperbolic(1:3) # accept the defaults
## Should return
# [1] 0.3333333 0.5000000 0.6000000
# attr(,"scale")
# [1] 2
}
\references{
The use of a variable hyperbolic transform to help model the
relations between stream water chemistry and flow was first described in:

Johnson, N.M., Likens, G.E., Borman, F.H., Fisher, D.W., and Pierce, R.S.,
1969, A working model for the variation in stream water chemistry at the
Hubbard Brook Experimental Forest, New Hampshire: Water Resources Research,
v. 5, no. 6, p. 1353-1363.
}
\seealso{
\code{\link{boxCox}}
}
\keyword{manip}

