% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/insertMissing.R
\name{insertMissing}
\alias{insertMissing}
\title{Insert Missing Data}
\usage{
insertMissing(x, col, fill = FALSE)
}
\arguments{
\item{x}{the data frame.}

\item{col}{the name of the column that defines the sequence.}

\item{fill}{logical (\code{TRUE} or \code{FALSE}), fill with many rows of data to
match the sequence?  If \code{FALSE}, insert only a single row. See \bold{Note}.}
}
\value{
A data frame like \code{x}, but with rows of missing values where
there is a break in the sequence in column \code{col}.
}
\description{
Inserts rows of missing values into a data frame for gaps in a sequence.
}
\note{
Setting \code{fill} to \code{TRUE} is useful for setting up datasets
for \code{fillMissing} if there are gaps in the retrieved data.  Setting
\code{fill} to \code{FALSE}, the default, is useful for creating a break
in the sequence for plotting the data.
}
\examples{
\dontrun{
library(smwrData)
data(Q05078470)
# Plot the original data
with(Q05078470[100:120, ], plot(DATES, FLOW, type="l"))
# Remove 3 rows from the data set
Q05078470 <- Q05078470[-(109:111), ]
# Plot the data--line drawn through the missing record
with(Q05078470[100:117, ], lines(DATES, FLOW, col="green"))
# Insert a missing record
Q05078470 <- insertMissing(Q05078470, "DATES")
# Now plot to show gap in line
with(Q05078470[100:118, ], lines(DATES, FLOW, col="blue"))
}
}
\seealso{
\code{\link{fillMissing}}, \code{\link{screenData}}
}
\keyword{manip}

