% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.na.timeDay.R
\name{is.na.timeDay}
\alias{is.na.timeDay}
\title{Missing Values}
\usage{
\method{is.na}{timeDay}(x)
}
\arguments{
\item{x}{the object to be tested.}
}
\value{
A logical vector of the same length as its argument \code{x},
containing \code{TRUE} for those elements marked \code{NA} and \code{FALSE} otherwise.
}
\description{
Indicate which elements are missing.
}
\examples{
is.na(as.timeDay(c("10:30", "11:00")))
}
\keyword{manip}
