% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.timeDay.R
\name{length.timeDay}
\alias{length.timeDay}
\title{Length of an Object}
\usage{
\method{length}{timeDay}(x)
}
\arguments{
\item{x}{a time-of-day object.}
}
\value{
An integer of length 1 indicating the number of elements in \code{x}.
}
\description{
Get the length of a time-of-day object.
}
\examples{

length(as.timeDay(c("10:30", "11:00")))
}
\keyword{attribute}
