% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makepredictcall.quadratic}
\alias{makepredictcall.boxCox}
\alias{makepredictcall.hyperbolic}
\alias{makepredictcall.quadratic}
\alias{makepredictcall.scaleRng}
\title{Utility Function for Safe Prediction}
\usage{
\method{makepredictcall}{quadratic}(var, call)

\method{makepredictcall}{hyperbolic}(var, call)

\method{makepredictcall}{boxCox}(var, call)

\method{makepredictcall}{scaleRng}(var, call)
}
\arguments{
\item{var}{a variable.}

\item{call}{the term in the formula, as a call.}
}
\value{
A replacement for \code{call} for the prediction variable.
}
\description{
A utility to help \code{\link{model.frame.default}} create the correct matrices
when predicting from models with \code{quadratic},
\code{hyperbolic}, or \code{boxCox} terms. Used only internally.
}

