\docType{methods}
\name{Math-methods}
\alias{Math,timeDay-method}
\alias{Math-methods}
\title{Mathematical Functions for \code{timeDay} objects}
\description{
  No mathematic functions, such as \code{log} or \code{exp}
  are allowed on time-of-day data. An error results when
  trying to use any mathematic function on objects of class
  "timeDay."
}
\section{Methods}{
  \describe{

  \item{signature(x = "timeDay")}{ Any call to a
  mathematical function will generate an error. } }
}
\keyword{methods}

