% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math-timeDay.R
\docType{methods}
\name{Math-timeDay}
\alias{Math,timeDay-method}
\alias{Math-timeDay}
\title{Mathematical Functions for \code{timeDay} objects}
\usage{
\S4method{Math}{timeDay}(x)
}
\arguments{
\item{x}{an object of class "timeDay."}
}
\description{
No mathematical functions, such as \code{log} or \code{exp} are allowed 
on time-of-day data. An error results when trying to use any mathematical function 
on objects of class "timeDay."
}
\keyword{methods}

