% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movingDiff.R
\name{movingDiff}
\alias{movingDiff}
\title{Moving Differences}
\usage{
movingDiff(x, span = 1, pos = "end")
}
\arguments{
\item{x}{the data to be averaged or differenced. Missing values are permitted
but result in missing values in the output.}

\item{span}{the span of the differences.}

\item{pos}{how to position the output data relative to the value returned;
"center" means that value represents the difference between the
preceding \code{span}/2 value and the following \code{span}/2 value, 
"end" or "trailing" means that value is the difference between the 
preceding \code{span} value and the current value, and "begin" or "leading" 
means that value is the difference between the current value and the 
following \code{span} value.}
}
\value{
A vector of the same length as \code{x} containing the differences.
}
\description{
Filter a regular series of data to compute a moving difference.
}
\examples{

# Construct a simple valley
movingData <- abs(seq(-5, 5))
movingDiff(movingData, span=1)
# Compare to diff:
diff(movingData)
}
\seealso{
\code{\link[stats]{filter}}, \code{\link[base]{diff}}, \code{\link{movingAve}}
}
\keyword{manip}

