% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na2miss.R
\name{na2miss}
\alias{na2miss}
\alias{miss2na}
\alias{miss2na}
\title{Recode Data}
\usage{
na2miss(x, to = -99999)

miss2na(x, from = -99999)
}
\arguments{
\item{x}{a vector. Missing values (\code{NA}s) are allowed.}

\item{to}{the replacement value for \code{NA}.}

\item{from}{the target value to match and replace with \code{NA}.}
}
\value{
An object like \code{x} with each target value replaced by the
specified value.
}
\description{
Converts missing values (\code{NA}s) to or from a user specified value.
}
\note{
The function \code{na2miss} converts missing values (\code{NA}) to the
value \code{to} and is useful to prepare a vector for export and subsequent
use by software external to R that does not handle NAs.\cr The function
\code{miss2na} converts the value \code{from} to \code{NA} and can be used to
recode data imported from external software that uses a special value to
indicate missing values.\cr
}
\examples{

## Construct simple substitutions
na2miss(c(1, 2, 3, NA, 5, 6))
}
\seealso{
\code{\link[base]{is.na}}, \code{\link[base]{sub}}
}
\keyword{manip}
