% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pick}
\alias{pick}
\title{Conditional Element Selection}
\usage{
pick(test, ..., .pass = test, na = NA)
}
\arguments{
\item{test}{a logical, numeric or character vector that indicates which value
to select from the data supplied in \code{\dots{}}. See \bold{Details}.}

\item{.pass}{the value to return for any element of \code{test} that does not
match an argument name in \dots{}. Useful only when the class of \code{test}
is "character" or "factor."}

\item{na}{the value to return for any element of \code{test} is \code{NA}.}

\item{\dots}{the values to be selected.}
}
\value{
A vector of the same length as \code{test} and data values from the
values list defined by \code{\dots{}}. The mode of the result will be coerced
from the values list defined by \code{\dots{}}.
}
\description{
Return the value associated with \code{test} from the supplied vectors.
}
\details{
If \code{test} is logical, then if test is \code{TRUE}, return the first
argument in \dots{}, otherwise return the second argument.\cr If \code{test}
is numeric, then return that value in the list defined by \code{\dots{}}.\cr
If \code{test} is character, then return that value in the list defined by
\code{\dots{}}, which must be named in the call.\cr If \code{test} is
\code{NA}, then return the value specified by \code{na}.
}
\note{
This function is designed to replace nested \code{ifelse} expressions.
See \bold{Examples}. It is different from \code{switch} in that the value
selected from the possible alternatives is selected by the values in
\code{test} rather than by a single value.
}
\examples{
## Create the test vector
testpick <- c(1,2,3,1)
## Nested ifelse
ifelse(testpick == 1, 1,
 ifelse(testpick == 2, 3,
   ifelse(testpick == 3, 5, NA)))
## Results by pick:
pick(testpick, 1, 3, 5)
## Create a test vector of character data
testpick <- c("a","b","c","a")
pick(testpick, a=1, b=3, c=5)
}
\seealso{
\code{\link{ifelse}}, \code{\link{switch}}
}
\keyword{manip}

