\name{print.timeDay}
\alias{print.timeDay}
\title{Print an Object}
\usage{
  \method{print}{timeDay}(x, ...)
}
\arguments{
  \item{x}{an object of class "timeDay."}

  \item{...}{not used, required for other methods.}
}
\value{
  The object \code{x} is returned invisibly
}
\description{
  Print objects defined in the USGSbase package.
}
\section{Side Effect}{
  The object \code{x} is printed
}
\examples{
\dontrun{}
}
\seealso{
  \code{\link{timeDay-class}}
}
\keyword{print}

