% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readList.R
\name{readList}
\alias{readList}
\title{Import Data}
\usage{
readList(file, names = TRUE, sep = "", nlines = 1, convert = NULL)
}
\arguments{
\item{file}{a character string specifying the name of the file.}

\item{names}{logical, if \code{TRUE}, then take component names from the first 
entry in the line. If \code{FALSE}, then the components are sequentially 
numbered.}

\item{sep}{the separator character for the data in each line. If a blank
string (the default), then any white space is taken as the separator.}

\item{nlines}{the number of lines that represent a single collection of data,}

\item{convert}{character string indicating how to convert the data. Must be
a valid value for the \code{Class} argument of \code{as}.}
}
\value{
A list with one component for each \code{nlines} in the input file.
}
\description{
Import data arranged on lines into a list.
}
\examples{
# Make a 3-line example dataset with component names A, B, and C.
cat("A 1 2 3 4\\nB 5 6 7\\nC 8 9\\n", file="readList.test")
# Read the example dataset
readList("readList.test")
}
\seealso{
\code{\link[methods]{as}}
}
\keyword{IO}
\keyword{manip}
