% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenData.R
\name{screenData}
\alias{screenData}
\title{Screen Data for Completeness}
\usage{
screenData(dates, values, type = "DV", year = "calendar", printit = TRUE)
}
\arguments{
\item{dates}{the sequence of dates for each value in \code{values}.}

\item{values}{the sequence of observations.}

\item{type}{the frequency of \code{values}. Only daily values ("DV") 
and intermittent, or discrete, values ("IV") are accepted in this version.
The whole text is required, but not case sensitive.}

\item{year}{the type of year: "calendar" or "water," which begins on October 1
of the previous calendar year and ends on September 30.}

\item{printit}{logical, if \code{TRUE}, then print the results in an}
}
\value{
For \code{type} = "DV," a matrix of the counts of \emph{missing} values,
either coded as \code{NA} or not in the dataset, for each month and each
year within the range of \code{dates}.

For \code{type} = "IV," a matrix of the counts of \emph{observed} values 
for each month and each year within the range of \code{dates}.
}
\description{
Screens data to determine if a value is reported for each date 
by calendar or water year.
}
\details{
Missing values are permitted in either \code{dates} or \code{values}. 
Those missing values are tallied in the completeness of record.
}
\examples{
library(smwrData)
data(Q05078770)
# this should indicate no missing values.
with(Q05078770, screenData(DATES, FLOW))
# There should be missing values shown for:
#months 10-12 in water year 2003 (October - December, 2002), and
#months 1-9 of water year 2004.  
with(Q05078770, screenData(DATES, FLOW, year="w"))
}
\references{
This function is based on the screen program described in:\cr
Rutledge, A.T., 2007, Program user guide for RECESS: \emph{at}
\url{http://water.usgs.gov/ogw/recess/UserManualRECESS.pdf}.
}
\keyword{missing}

