% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setFileType.R
\name{setFileType}
\alias{setFileType}
\title{File or Object Name}
\usage{
setFileType(filename, type = "tmp", replace = FALSE)
}
\arguments{
\item{filename}{a single character string of the name of the file or object.}

\item{type}{character string identifying the new suffix.}

\item{replace}{logical (\code{TRUE} or \code{FALSE}): replace the current suffix?}
}
\value{
A character string like \code{filename} but with a new suffix.
}
\description{
Create a new file or object name from an old name with an optional new
suffix.
}
\note{
This function is designed as a support function for many
functions.
}
\examples{
# Replace the .dat suffix with .txt
setFileType("TestName.dat", "txt", replace=TRUE)
}
\author{
Dave Lorenz, original coding by Jim Slack, U.S. Geological Survey retired.
}
\keyword{MANIP}

