\docType{methods}
\name{show-methods}
\alias{show-methods}
\alias{show,timeDay-method}
\title{Methods for Function \code{show} for time-of-day objects}
\description{
  Display the time of day.
}
\section{Methods}{
  \describe{

  \item{list("signature(object = \"timeDay\")")}{ Display
  to time of day in the default format. } }
}
\keyword{~~}
\keyword{keyword(s)}
\keyword{methods}
\keyword{other}
\keyword{possible}

