\docType{class}
\name{timeDay-class}
\alias{timeDay-class}
\title{timeDay class}
\description{
  The timeDay class is a simple representation of the time
  of day. The data are stores as seconds since midnight.

  Class "timeDay" describes the time of day without any
  reference to the date.
}
\section{Objects from the Class}{
  Objects can be created by calls of the form
  \code{as.timeDay(time, format)}.
}
\references{
  Lorenz, D.L., 2013, The USGSwsBase package
}
\keyword{classes}

