\docType{package}
\name{USGSwsBase-package}
\alias{USGSwsBase}
\alias{USGSwsBase-package}
\title{Data import, export and manipulation functions}
\description{
  This package has specialized functions for importing,
  managing, or manipulating hydrologic data.
}
\details{
  \tabular{ll}{ Package: \tab USGSwsBase\cr Type: \tab
  Package\cr Version: \tab 0.8.0\cr Date: \tab
  2014-02-06\cr License: \tab File LICENSE\cr Depends: \tab
  stat, lubridate, methods\cr } This package contains
  functions that import and export data from/to a format
  supported by the National Water Information System (NWIS)
  and the web access version (NWISweb). It also has
  functions that manage or manipipulate hydrologic data and
  functions that apply specialized transforms.
}
\author{
  Dave Lorenz <lorenz@usgs.gov>

  Maintainer: Dave Lorenz <lorenz@usgs.gov>
}
\references{
  Lorenz, D.L., in preparation, USGSwsBase---an R package
  for managing hydrologic data, version 0.8.0: U.S.
  Geological Survey Open File Report 2014-XXXX, ?p.
}
\seealso{
  \code{\link[USGSwsData:USGSwsData-package]{USGSwsData}}
}
\keyword{package}

