\name{as.character.timeDay}
\alias{as.character.timeDay}
\title{Character Vector}
\usage{
  \method{as.character}{timeDay}(x, ...)
}
\arguments{
  \item{x}{the object ot be converted.}

  \item{\dots}{further arguments passed to or from other
  methods.}
}
\value{
  The values in \code{x} converted to a character
  representation.
}
\description{
  Convert to object of type "character."
}
\keyword{character}

