% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.character.timeDay}
\alias{as.character.timeDay}
\title{Character Vector}
\usage{
\method{as.character}{timeDay}(x, ...)
}
\arguments{
\item{x}{the time-of-day object to be converted.}

\item{\dots}{not used, required for other methods.}
}
\value{
The values in \code{x} converted to a character representation.
}
\description{
Convert the time-of-day object to a character string.
}
\keyword{character}

