\name{c.timeDay}
\alias{c.timeDay}
\title{Concatenate Data}
\usage{
  \method{c}{timeDay}(..., recursive = FALSE)
}
\arguments{
  \item{\dots}{any number of objects that can be converted
  to class "timeDay."}

  \item{recursive}{required for other methods.}
}
\value{
  A single vector of class "timeDay."
}
\description{
  Combine data into a single vector.
}
\keyword{manip}

