% Generated by roxygen2 (4.0.2): do not edit by hand
\name{conc.meq}
\alias{conc.meq}
\title{Special Data}
\usage{
conc.meq()
}
\value{
A list containing necessary information for the function
\code{conc2meq}.
}
\description{
Support function to supply data for functions within smwrBase.
}
\note{
The user may choose to make local copies of the data with the
same name (conc.meq) to be able to change or add to the list.
}
\seealso{
\code{\link{conc2meq}}
}
\keyword{list}

