% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importCSV.R
\name{importCSV}
\alias{importCSV}
\title{Import Files}
\usage{
importCSV(file.name = "", tz = "")
}
\arguments{
\item{file.name}{a character string specifying the name of the comma-separated variable 
(CSV) file containing the data to be imported; \code{importCSV} requires \code{file.name}
to be a readable file on the computer.}

\item{tz}{a character string indicating the time-zone information for data
imported as "POSIXct." The default is to use the local setting.}
}
\value{
A data frame with one column for each data column in the CSV
file.
}
\description{
Imports a comma-separated variable file to a data frame.
}
\details{
All of the dates in a date column must have the same format as the first
non-blank date in the column. Any date with a format different from that of
the first non-blank date in the column will be imported as \code{NA} (missing
value). Dates imported as class "Date" using a 4-digit year,
2-digit month, and 2-digit day with the period (.), hyphen (-), slash (/), or
no separator. Time and date data are imported as class "POSIXct" and
assumes the standard POSIX format for date and time.\cr
}
\note{
A NULL data frame is created if there are no data in the file.\cr
}
\examples{
\dontrun{
## These datasets are available in smwrData as text files
TestDir <- system.file("misc", package="smwrData")
TestPart <- importCSV(file.path(TestDir, "TestPart.csv"))
}
}
\seealso{
\code{\link[utils]{read.csv}}, \code{\link[utils]{read.table}}, \code{\link[base]{scan}},
\code{\link[base]{as.Date}}, \code{\link[base]{as.POSIXct}}
}
\keyword{IO}
\keyword{manip}
