% Generated by roxygen2 (4.0.2): do not edit by hand
\name{insertMissing}
\alias{insertMissing}
\title{Insert Missing Data}
\usage{
insertMissing(x, col, fill = FALSE)
}
\arguments{
\item{x}{the data frame.}

\item{col}{the name of the column that defines the sequence.}

\item{fill}{logical (\code{TRUE} or \code{FALSE}), fill with many rows of data to
match the sequence?  If \code{FALSE}, insert only a single row. See \bold{Note}.}
}
\value{
A data frame like \code{x}, but with rows of missing values where
there is a break in the sequence in column \code{col}.
}
\description{
Inserts rows of missing values into a data frame for gaps in a sequence.
}
\note{
Setting \code{fill} to \code{TRUE} is useful for setting up datasets
for \code{fillMissing} if there are gaps in the retrieved data.  Setting
\code{fill} to \code{FALSE}, the default, is useful for creating a break
in the sequence for plotting the data.
}
\seealso{
\code{\link{fillMissing}}, \code{\link{screenData}}
}
\keyword{manip}

