\docType{methods}
\name{Math-methods}
\alias{Math-methods}
\alias{Math,timeDay-method}
\title{Mathematic Methods for \code{timeDay} objects}
\description{
  No mathematic methods are allowed on time-of-day data. An
  error results from trying to use any mathematic function
  on objects of class "timeDay."
}
\section{Methods}{
  \describe{

  \item{list("signature(x = \"timeDay\")")}{ Any call to a
  mathematical function will generate an error. } }
}
\keyword{methods}

