% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/more.R
\name{more}
\alias{more}
\title{Display Data}
\usage{
more(x, n = 20L, ...)
}
\arguments{
\item{x}{any valid object, generally a data frame, matrix, or table.}

\item{n}{a positive integer indicating how many lines to print for a page.}

\item{\dots}{additional arguments to be passed to methods for \code{head} or
\code{tail}.}
}
\value{
The object \code{x} is returned invisibly. Sections of \code{x} of
length \code{n} are displayed during the execution of the function.
}
\description{
Display the contents of an object by pages.
}
\note{
The function \code{more} is intended for interactive sessions. If used
in a non-interactive session, it simply returns \code{x} invisibly.

Several keyboard commands can be used to view the contents of \code{x}. The 
function \code{more} will display \code{n} lines of \code{x} and wait for 
user input. Any of the following commands can be entered by the user; 
either upper- or lowercase letters are accepted.

\code{q} Quit

\code{t} Go to top of \code{x}

\code{b} Go to bottom of \code{x}

\code{u} Go up 1/2 page

\code{p} Go to previous page

\code{d} Go down 1/2 page

\code{colName/pattern} Search for pattern in the column named colName

\code{h} or \code{?} Print help

\code{any other letter} Go down full page

Searching for a pattern in a column uses \code{grep} to search for the
specified pattern in the character representation of the data in the column.
This makes it possible to search columns that are not type character.
}
\seealso{
\code{\link[utils]{head}}, \code{\link[utils]{tail}}, \code{\link[base]{grep}}
}
\keyword{print}

