% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.timeDay.R
\name{print.timeDay}
\alias{print.timeDay}
\title{Print an Object}
\usage{
\method{print}{timeDay}(x, ...)
}
\arguments{
\item{x}{an object of class "timeDay."}

\item{\dots}{not used, required for other methods.}
}
\value{
The object \code{x} is returned invisibly.
}
\description{
Print an object of class "timeDay."
}
\note{
The object is printed using the format that created the object
in \code{as.timeDay}.
}
\section{Side Effect}{
 The object \code{x} is printed.
}

\seealso{
\code{\link{timeDay-class}}, \code{\link{as.timeDay}}
}
\keyword{print}
