% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleRng.R
\name{scaleRng}
\alias{scaleRng}
\alias{IscaleRng}
\alias{IscaleRng}
\title{Scale Data}
\usage{
scaleRng(x, Min = 0, Max = 1, x.range = range(x, na.rm = TRUE))

IscaleRng(x, Min, Max, x.range)
}
\arguments{
\item{x}{any numeric vector. Missing values are permitted and
result in missing values in the corresponding output.}

\item{Min}{the minimum of the output range.}

\item{Max}{the maximum of the output range.}

\item{x.range}{the input range to map to the output range. The default range is 
computed from the range of \code{x} after removing missing values.}
}
\value{
A numeric vector scaled to the specified range.
}
\description{
Transforms numeric data to a specified range.
}
\details{
The function \code{scaleRng} maps the minimum of \code{x.range} to \code{Min}
and the maximum of \code{x.range} to \code{Max} and uses linear interpolation 
for other values in \code{x}.
}
\note{
Some applications recommend or require data scaled to a consistent range.
The function \code{scaleRng} will do that and can be used to back-transform
the data.
}
\examples{

## simple case with back-transform
x.tmp <- print(scaleRng(c(1.2, 2.3, 3.4, 5.6)))
IscaleRng(x.tmp)
## now set the expected ranges
x.tmp <- print(scaleRng(c(1.2, 2.3, 3.4, 5.6), x.range=c(1, 6)))
IscaleRng(x.tmp)

}
\keyword{manip}
