\docType{methods}
\name{show-methods}
\alias{show,timeDay-method}
\alias{show-methods}
\title{Methods for Function \code{show} for time-of-day objects}
\arguments{
  \item{object}{the object to be printed.}
}
\value{
  The object is printed and returned invisibly.
}
\description{
  Display the time of day.
}
\keyword{methods}

