% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{show-methods}
\alias{show,timeDay-method}
\alias{show-methods}
\title{Methods for Function \code{show} for Time-of-Day Objects}
\usage{
\S4method{show}{timeDay}(object)
}
\arguments{
\item{object}{the object to be printed.}
}
\value{
The object is printed and returned invisibly.
}
\description{
Display the time of day.
}
\keyword{methods}

