\docType{class}
\name{timeDay-class}
\alias{timeDay-class}
\title{timeDay class}
\description{
  Some details about the timeDay class

  Class "timeDay" describes the time of day without any
  reference to the date.
}
\section{Objects from the Class}{
  Objects can be created by calls of the form
  \code{as.timeDay(time, format)}.
}
\references{
  Lorenz, D.L., 2013, The USGSwsBase package
}
\keyword{classes}

