% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeDay-class.R
\docType{class}
\name{timeDay-class}
\alias{timeDay-class}
\title{Time of Day}
\description{
Class "timeDay" describes the time of day without any reference to the date.
The data are stored as seconds since midnight.
}
\section{Slots}{

\describe{
\item{\code{time}}{the time of day in seconds since midnight.}

\item{\code{format}}{a character string indicating the format to display the time of day.}
}}

\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{as.timeDay(time, format)}.
}

\examples{

showClass("timeDay")

}
\keyword{classes}
