% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z[.timeDay.R
\name{[.timeDay}
\alias{[.timeDay}
\title{Extract Parts of an Object}
\usage{
\method{[}{timeDay}(x, i)
}
\arguments{
\item{x}{the object.}

\item{i}{an index specifying elements to extract. See \code{\link{Extract}}
for details.}
}
\value{
The subset of \code{x} indicated by \code{i}.
}
\description{
Extract elements of a time-of-day object.
}
\seealso{
\code{\link[base]{Extract}}
}
\keyword{manip}

